Move extension methods to the interim package
