+++ core/src/main/scala/eu/joaocosta/interim/api/Components.scala
@@ -1,6 +1,5 @@
 package eu.joaocosta.interim.api
 
-import eu.joaocosta.interim.ItemId.*
 import eu.joaocosta.interim.*
 import eu.joaocosta.interim.skins.*
